+++ lib/src/custom_clipper.dart
@@ -9,9 +9,9 @@ class HalfClipper extends CustomClipper<Rect> {
   /// Create a [Rect] on widget ViewPort to clip the widget
   /// based on [clip] value
   @override
-  Rect getClip(Size size) => Rect.fromLTRB(0, 0, size.width * clip, size.height);
+  Rect getClip(Size size) =>
+      Rect.fromLTRB(0, 0, size.width * clip, size.height);
 
-    @override
-    bool shouldReclip(covariant CustomClipper<Rect> oldClipper) => true;
-
+  @override
+  bool shouldReclip(covariant CustomClipper<Rect> oldClipper) => true;
 }