+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.2.2 - [29/01/2022]
+
+* Added IconScoreBar that can used with custom icons.
+* Added onChanged method to get notified when the score value is updated.
+
 ## 0.0.1 - [24/01/2022]
 
-* Added ScoreBar that can be mutable/immutable
+* Added ScoreBar that can be mutable/immutable.